librbd: ssize_t return values for read, write
size_t is 32bits on 64bit archs.  Use ssize_t (long) for return values.

+++ src/librbd.cc
@@ -133,7 +133,7 @@ namespace librbd {
     Mutex lock;
     Cond cond;
     bool done;
-    int rval;
+    ssize_t rval;
     callback_t complete_cb;
     void *complete_arg;
     rbd_completion_t rbd_comp;
@@ -171,9 +171,9 @@ namespace librbd {
 
     void complete_block(AioBlockCompletion *block_completion, int r);
 
-    int get_return_value() {
+    ssize_t get_return_value() {
       lock.Lock();
-      int r = rval;
+      ssize_t r = rval;
       lock.Unlock();
       return r;
     }
@@ -251,8 +251,8 @@ namespace librbd {
   int64_t read_iterate(ImageCtx *ictx, uint64_t off, size_t len,
 		       int (*cb)(uint64_t, size_t, const char *, void *),
 		       void *arg);
-  int read(ImageCtx *ictx, uint64_t off, size_t len, char *buf);
-  int write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf);
+  ssize_t read(ImageCtx *ictx, uint64_t off, size_t len, char *buf);
+  ssize_t write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf);
   int aio_write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf,
                 AioCompletion *c);
   int aio_read(ImageCtx *ictx, uint64_t off, size_t len,
@@ -1097,12 +1097,12 @@ static int simple_read_cb(uint64_t ofs, size_t len, const char *buf, void *arg)
 }
 
 
-int read(ImageCtx *ictx, uint64_t ofs, size_t len, char *buf)
+ssize_t read(ImageCtx *ictx, uint64_t ofs, size_t len, char *buf)
 {
   return read_iterate(ictx, ofs, len, simple_read_cb, buf);
 }
 
-int write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf)
+ssize_t write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf)
 {
   if (!len)
     return 0;
@@ -1111,7 +1111,7 @@ int write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf)
   if (r < 0)
     return r;
 
-  int total_write = 0;
+  size_t total_write = 0;
   uint64_t start_block = get_block_num(&ictx->header, off);
   uint64_t end_block = get_block_num(&ictx->header, off + len - 1);
   uint64_t block_size = get_block_size(&ictx->header);
@@ -1218,7 +1218,7 @@ int aio_write(ImageCtx *ictx, uint64_t off, size_t len, const char *buf,
   if (r < 0)
     return r;
 
-  int total_write = 0;
+  size_t total_write = 0;
   uint64_t start_block = get_block_num(&ictx->header, off);
   uint64_t end_block = get_block_num(&ictx->header, off + len - 1);
   uint64_t block_size = get_block_size(&ictx->header);
@@ -1382,7 +1382,7 @@ int RBD::AioCompletion::wait_for_complete()
   return c->wait_for_complete();
 }
 
-int RBD::AioCompletion::get_return_value()
+ssize_t RBD::AioCompletion::get_return_value()
 {
   librbd::AioCompletion *c = (librbd::AioCompletion *)pc;
   return c->get_return_value();
@@ -1458,7 +1458,7 @@ int Image::snap_set(const char *snap_name)
   return librbd::snap_set(ictx, snap_name);
 }
 
-int Image::read(uint64_t ofs, size_t len, bufferlist& bl)
+ssize_t Image::read(uint64_t ofs, size_t len, bufferlist& bl)
 {
   ImageCtx *ictx = (ImageCtx *)ctx;
   bufferptr ptr(len);
@@ -1473,7 +1473,7 @@ int64_t Image::read_iterate(uint64_t ofs, size_t len,
   return librbd::read_iterate(ictx, ofs, len, cb, arg);
 }
 
-int Image::write(uint64_t ofs, size_t len, bufferlist& bl)
+ssize_t Image::write(uint64_t ofs, size_t len, bufferlist& bl)
 {
   ImageCtx *ictx = (ImageCtx *)ctx;
   if (bl.length() < len)
@@ -1671,7 +1671,7 @@ extern "C" int rbd_snap_set(rbd_image_t image, const char *snapname)
 }
 
 /* I/O */
-extern "C" int rbd_read(rbd_image_t image, uint64_t ofs, size_t len, char *buf)
+extern "C" ssize_t rbd_read(rbd_image_t image, uint64_t ofs, size_t len, char *buf)
 {
   librbd::ImageCtx *ictx = (librbd::ImageCtx *)image;
   return librbd::read(ictx, ofs, len, buf);
@@ -1684,7 +1684,7 @@ extern "C" int64_t rbd_read_iterate(rbd_image_t image, uint64_t ofs, size_t len,
   return librbd::read_iterate(ictx, ofs, len, cb, arg);
 }
 
-extern "C" int rbd_write(rbd_image_t image, uint64_t ofs, size_t len, const char *buf)
+extern "C" ssize_t rbd_write(rbd_image_t image, uint64_t ofs, size_t len, const char *buf)
 {
   librbd::ImageCtx *ictx = (librbd::ImageCtx *)image;
   return librbd::write(ictx, ofs, len, buf);
@@ -1717,7 +1717,7 @@ extern "C" int rbd_aio_wait_for_complete(rbd_completion_t c)
   return comp->wait_for_complete();
 }
 
-extern "C" int rbd_aio_get_return_value(rbd_completion_t c)
+extern "C" ssize_t rbd_aio_get_return_value(rbd_completion_t c)
 {
   librbd::RBD::AioCompletion *comp = (librbd::RBD::AioCompletion *)c;
   return comp->get_return_value();
